+++ docs/root/intro/version_history.rst
@@ -5,6 +5,7 @@ Version history
 ===============
 * admin: added :http:get:`/hystrix_event_stream` as an endpoint for monitoring envoy's statistics 
   through `Hystrix dashboard <https://github.com/Netflix-Skunkworks/hystrix-dashboard/wiki>`_.
+* http: better handling of HEAD requests. Now sending transfer-encoding: chunked rather than content-length: 0
 * http: response filters not applied to early error paths such as http_parser generated 400s.
 * ratelimit: added support for :repo:`api/envoy/service/ratelimit/v2/rls.proto`.
   Lyft's reference implementation of the `ratelimit <https://github.com/lyft/ratelimit>`_ service also supports the data-plane-api proto as of v1.1.0.